REFACTOR SignIn and SignUp pages with UserServices
